+++ Race/src/main/java/model/air/MagicCarpet.java
@@ -6,7 +6,7 @@ public class MagicCarpet extends AirVehicle {
     }
 
     @Override
-    public int getDistanceReducer(int length) {
+    public float getDistanceReducer(float length) {
         if(length < 1000) return 0;
         else if(length < 5000) return 3;
         else if(length < 10000) return 10;
